move root-level Rust files into their appropriate parent modules
